CAMERA SCRIPT WORKING
I had ChatGPT annotate the code if anyone needs to understand it. We basically just create a webcam object attached to a raw image UI object. The button runs the script to capture an image and save it to the newly created ImageCaptures folder. The path should be universal but let me know if it is not. Thanks. Next step: make the button save the image and then take them to a results screen (info).