a lot of changes
finalized (for the most part) a lot of the backend code and added docstrings (with the exception of the algorithms file, but I'm too tired to do that rn). refactored a lot of files in order to make the folder make more sense. removed the visuals file, and put all those functions under utils.